refactor: extract transfer business logic from commands/ to core/
Move Target enum, parse_target(), SSH arg builders, and all transfer
execution functions (push/pull/server-to-server/scp) to core/transfer.rs.

commands/transfer.rs is now a thin CLI layer: clap args → TransferConfig → core::transfer().

493 → 49 lines in commands layer. Zero behavior change.
